Ajustes finos + Inclusão de logs.

+++ src/main/java/com/example/demo/controller/dto/EntregaRepostaDTO.java
@@ -3,6 +3,6 @@ package com.example.demo.controller.dto;
 public record EntregaRepostaDTO(
         Long id,
         String descricao,
-        byte nota
+        Double nota
 ) {
 }
